The sms lexicon is rewritten into Xerox conventions by Lauri Karttunen, based up\on the Lexicon in the Trosterud / Moshagen NJL article.
Multichar symbols simply lists all multichar symbols.

The Root lexicon contains the Noun sublexicon only.

The first 30 or so entries of the Noun sublexicon display all the
declension types. Then follows, in alphabetical order, all nominal
root entries of the Bergsland / Magga dictionary. They were taken from
the original Lingosft format and translated into Xerox format. At the
end of the file comes the morphological sublexica.

The good thing about the Xerox format is that it makes it possible to
generate the sms paradigms. The bad thing is that it contains
errors. The gravest one is that it does not handle the syllable
counting that is needed to ensure suffix alternation (sne / snie,
etc.). This should be fixed.

TODO: The file should be split into two files, in the same way as the
sme files.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sma@22 c7155fb1-f0a7-4240-a2fc-2600b6f42f90