Now made own lexica for u- and o-final names, they were all wrong in the present VNAME system. All nominatives had an -e added, etc.
The first migrate to UNAME, keyforms being Timbuktu, Timbuktuvem, Timbuktuvesne.
Still to be fixed here is the even- and odd-syllabic patterns.

The second migrate to ONAME, keyforms Aino, Ainoesne.
Note that Aino gets both Ainom and Ainoem in accusative, the latter in order to
align it with oe-words. She still does not get a final -e in nominative (*Ainoe).
What to do here will have to be decided by normative bodies (and corpus research).


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sma@27651 c7155fb1-f0a7-4240-a2fc-2600b6f42f90