This is the bïssedh to bissien (with i) fix.
The rule
"Even syllabic verbs Du1Pl3 IV"
was changed to have its right context e optional, in order to allow for bïssedh.
The new version of the actual line is here, with (e) for former e:
           [ Cns (%»: %^adteUML:) | #: ]       _ (e) Cns+ [ %>: | %»: ] %^3dUML: ;           
If we had run our tests we would have noticed the error earlier, as the rule
was accompagnied with a Du1 version of mïnnedh -> minnien.