Added final -e to some NIEJTE lemma (novell, arovm, ...)
Removed final -e to some NIEJTE stems (etaate, ...)

Cosmetics: space before ;, but not before newline, for uniform testing.
Also squeezed multiple spaces.

A class of oe nouns had MAANA as contlex. I took it that the lemma was correct, and changed accordingly. Commented out as OE-MAANA, for eventual reference.

It could be that this was intended, in which case they should be put in again.
In that case I would suggest lemma in -a, so that the nominative of "hammoe" is not "hamma".


Changed the contlex 


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sma@36777 c7155fb1-f0a7-4240-a2fc-2600b6f42f90