Removed Adp, we now use Pr and Po. Also removed the Adp rules, we have Pr and Po rules.
Added some sets from sme, INSTITUTION, EDUCATION, ORGANIZATION, MEDIA.
These are used for the Prop disambiguation set, which I somewhat
tentatively glued in from sme, and adjusted.

Rewrote the Interj rule so that it works as intended:
Keep if the only word in the sentence, remove elsewhere. This is
an oversimplification, but will do as long as our corpora are
so narrow.

Went through the whole file and removed rules not in use. Now almost all the
rules (ehem, perhaps save the newly added sme-Prop ones) are in use,
and also some sensible but unused rules were kept, such as the ones
referring to Du2 and Pl2 (underrepresented in our corpus).

Added some word-specific rules (almetje, Jupmele, vihth, meatan).