This is a major rework: 389 lines were added, 383 were removed.
The starting point for the work was the decision to change from +Pred to
+Sg+Nom. At the outset, this is a technical search-replace-unify operation 
only, as the two tags are two names for the same thing. In practice, the sets
of all +Pred forms and of all +Sg+Nom forms were not identical. In meeting
this challenge I have aimed at a union operation: All forms tagged as
+Pred, +Sg+Nom, or both, were included in the new list. 

Unfortunately, this was initiated at the same time as we initiated a process 
of correcting adjective comparative and superlative forms. These corrections
were added to this file in parallel to the svn form.

Very many of the corrections were removing comparatives and superlatives, other
were adding of stem vowel for the comparative forms (credit to Maja for this
work). 

Some loanword lexica (such as AALE) were rewritten to open for the substandard
way of writing wit single a.

As far as possible, comparative and superlative forms were avoided as
baseforms.

Also, the file was sorted.