This is a major rework, 188 lines were added, 144 were removed.
The starting point for the work was the decision to change from +Pred to
+Sg+Nom. At the outset, this is a technical search-replace-unify operation 
only, as the two tags are two names for the same thing. In practice, the sets
of all +Pred forms and of all +Sg+Nom forms were not identical. In meeting
this challenge I have aimed at a union operation: All forms tagged as
+Pred, +Sg+Nom, or both, were included in the new list. 

Many of the case lexica were split in a Sg Nom lexicon, and an oblique
case lexicon. In this way, the variation in the Sg Nom (former Pred)
form could be kept in isolation from the case paradigm itself.

Several loan word paradigms (AALE, AARE, EETE, AANE..) were split in
order to allow from ale, are, ete, ane subforms.

The former doubling of contlex pairs like PRED-0 and eCASE was removed. Now
only one remains.

At the same time, we made a revision of the comparative and superlative
forms of the adjectives. In order to cope with this, I added more NOCOMP
lexica, and split the case lexica into positive and comparative/superlative
parts. In that way, one might point to one, the other, or both.

The lexica LOHKIJE, BÅETIJE were revised according to last week's meeting 
on these forms.

All this in one revision is clearly too much, but the reason for this
was that the process came at the same time as the adjective revision.
Now doubt errors have crept into this revision, that were not present in 
the previous one. On the other hand, this version is closer to the gold
standard than the previous one, so it now seems more productive to correct
this one than to carry on with the old.