A lot of whitespace changes to make it easier to sync sma with smj and sme, and at the same time make it easier to debug and change tags. Moved comments around.
Made numerals part of the R lexicons, that is, the regular compounding (requiring hyphens of course).

Added the +Cmp tag as part of all compounds (cf the Apertium change in sme and smj from last week).

Renamed +Cmpnd to +RCmpnd to sync with the other languages and make the semantics of the tag clearer. Renamed +Cmpnd+ to +LCmpnd for the same reason.

Moved all word boundaries in the R lexicons last in the symbol string - this is more logical, as the tags and the flag diacriticsbelong to the word to the left of the word boundary #. Now all tags and multichar symbols are on the same side of the # as the word itself.

Added +Der tag, even though it is not needed in sma (there are no non-positional derivations so far). By having it defined the languages are more in sync, and if we need non-positional derivations also for sma it is already there so we won't risk that it is not defined.

Added +Multi from smj.

More comments to make the purpose of each section clearer.