Changed the rule "a/e alternation".
It used to be
!   a:e <=>         _ [ Cns - j | %»: t j | h [ t | k ] ] (Cns) (e) (%>:) %^DISIMP: ;
but is now
   a:e <=>         _  Cns+ (e) (%>:) %^DISIMP: ;

The thing is: The rule as it stood restricted the vowel simplification rule 
to stems with only certain consonant clusters. Cf. Bergsland: "om
konsonanten mellom annen stavelse og tredje stavelse er j, er den
korte vokal i annan stavelse i eller o" (p. 105.)

But our starting point is the opposite. We know the underlying
vowel is -a-, and we know it will be reduced to -e- whenever there
is the DISIMP mark. Hence, it seems to me, we do not need the 
specification in the consonant cluster.



git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sma@59901 c7155fb1-f0a7-4240-a2fc-2600b6f42f90