This is the gemtebe fix.
The rule
  "a/e alternation"
has been restored to its original form, from before I changed it:

"a/e alternation"
 a:e <=> _ [ Cns - j | %»: t j | h [ t | k ] ] (Cns) (e) (%>:) %^DISIMP: ;
         _                            (%>:)     %^COMPDISIMP: [b:|m:|l:] ;
 [e:|u:] _   Cns+      %»:                            %^didhUML: e [d|l] ;

I still think my second outcommented line (see file) is better, linguistically
speaking, but I must of course give in to the test results.

As for test results:

It is sort of hard to test the sma analyser at its present stage:
When I run the adjective test in sma/test/morphology/,
I get 51 missing adjectives with the restored rule (this checkin), 
but 3117 missing adjectives with the version preceeding this.
These numbers are probably not real, but reflecting the present
stage of the testing files rather than the condition of the lexc and
twolc files.

I did, however, test the rule against the gold standard, revision 59359.
For that revision, and for testadj.sh -r59876, I got one missing
adjective only, njaakedihks.

Also, the present checkin gives us gamte - gemtebe, as required.
But whether it has introduced 51 new errors or not I am unfortunately
not able to tell. At least, it can easily be reversed.