Reorganised the test dir, in anticipation of a larger set of tools and source types in need of testing. Merge from the template.
Corrected the paths in the test shell scripts as well, so the tests are working with the reorganized structure.

NB! Please note that the new structure requires that you run ./autogen.sh + ./configure before running make/make check!!! This is needed to update the generated makefiles, and their knowledge of the new structure.