Speller test results using the latest nightly build from yesterday, with zhfst file from Tino. For a very strange reason hfst-ospell-office crashes when using zhfst files built on my systems, Tino is looking into it.
Results: more false negatives, which lowers recall and precision. On average slightly better position of the correct suggestions, and a bit less noise (fewer irrelevant suggestions).

MUCH slower, which is something we need to look into. Earlier SMA was fast, now it is acceptable, but there is a handful of words triggering suggestion times of more than one second (sort on the time column to easily find them).

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sma@160643 c7155fb1-f0a7-4240-a2fc-2600b6f42f90