Another attempt at playing with demoting edits to the first letter. This version compiles extremely fast, so that is not an issue. But the results...
Average # of suggs: 9.99

% correct #1:       7.71
% correct ≥5:      11.23
% correct any #:   13.08
% no sugg:          0.00
% only bad sug:    86.92

That is, suggestion-wise this version is PURE CRAP! Also, the speed was hillarious (0.18 words pr second, with a huge part taking ~6 seconds to generate suggestions). Obviously it is not enabled.

General insight so far: do not try to target the first letter specifically, it just ruins the speller. A better strategy would most likely be to target edits in non-initial positions and promote them (ie give them lower weight than default).